[ARM] Disable Newlib supplied syscalls in order to avoid SW breaks on stm32f4-discovery.
The star files of the stm32f4-discovery board have some optional software
breaks that are inserted if we are using newlib syscalls.
In order to be able to test the board with embench without these breaks
we need to prevent the compiler from using such newlib syscalls.

Files changed:

        * build_toolchain: add the --disable-newlib-supplied-syscalls option
          to the configuration of Newlib.