[NEWLIB-NANO] Add support to build newlib nano.
This differs from normal newlib for a mere matter of
configuration.
The configuration we had before for newlib was pretty close
to what is used for newlib-nano.
We add here the same configuration for newlib nano and the
creation of the nano* copies of the library files.
It will be up to a later change to give newlib a standard
configuration that differs from newlib nano.